Return QNetworkReply from getRequests
This patch allows the client implementation to react to specific
getRequests, such as folder checks, etc. It allows to connect
client-specific slots to requests, catch its errors and handle them
separately.

I need this patch to be able to check more than one remote folder on the
server, and not mix up their job's signal handling.