Add more debug in case inotify fails
We should probably make this at least a visible error, because it's a
limitation set by the kernel/the distro and we would require root
permissions to adjust the value.

Right now all that happens is that syncing is not immediate, but only
run every 30 secs (due to server polling).