make BUILD_WITH_QT4 off by default
Rationale:
When Qt5 is not found, the script will automatically use Qt4.
However, without this change, when Qt4 is not found, it will not
try to find Qt5 because BUILD_WITH_QT4 is ON by default forbidding the
use of Qt5 even if it is installed.

Also, other packages default to Qt5 including qtkeychain

See also issue #1418