SqlQuery: Reset after each successful exec().
In e5269a4151799ac442a47e011926feb98d7e5caa I moved the reset calls to
before the first bindValue() because we didn't consistently reset()
after exec().

However, some queries may keep a lock on the database so it is necessary
to reset() as soon as the data is read. This commit reintroduces these
reset() calls.